+++ src/Database/QueryBuilder/PgsqlQueryBuilder.php
@@ -231,7 +231,7 @@ class PgsqlQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterf
         } elseif ($column->getType() === Column::TYPE_TIMESTAMP && $column->getSettings()->getDefault() === ColumnSettings::DEFAULT_VALUE_CURRENT_TIMESTAMP) {
             $default = 'CURRENT_TIMESTAMP';
         } else {
-            $default = "'" . str_replace("'", "''", $column->getSettings()->getDefault()) . "'";
+            $default = "'" . $this->sanitizeSingleQuote($column->getSettings()->getDefault()) . "'";
         }
 
         return $default;
@@ -331,13 +331,18 @@ class PgsqlQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterf
 
     private function createTableComment(MigrationTable $table): string
     {
-        $comment = str_replace("'", "''", (string)$table->getComment());
+        $comment = $this->sanitizeSingleQuote((string)$table->getComment());
         return "COMMENT ON TABLE {$table->getName()} IS '$comment';";
     }
 
     private function createColumnComment(MigrationTable $table, Column $column): string
     {
-        $comment = str_replace("'", "''", (string)$column->getSettings()->getComment());
+        $comment = $this->sanitizeSingleQuote((string)$column->getSettings()->getComment());
         return "COMMENT ON COLUMN {$table->getName()}.{$column->getName()} IS '$comment';";
+    }
+
+    private function sanitizeSingleQuote(string $input): string
+    {
+        return str_replace("'", "''", $input);
     }
 }